refactor(renderer): refactor render single field
